+++ packages/web-tokens/CHANGELOG.md
@@ -2,6 +2,33 @@
 
 The following is a curated list of changes in the web tokens module, newest changes on the top.
 
+## [0.9.0] - 2025-11-07
+
+### Removed
+
+- `semantic-color-surface-badge-dark`
+- `semantic-color-surface-badge-light`
+- `semantic-color-on-surface-default-error`
+- `semantic-color-on-surface-default-success`
+- `semantic-color-on-surface-button`
+
+### Changed
+
+- `semantic-color-background-main` value
+- `semantic-color-on-background-default` value
+- `semantic-color-surface-default-handle` value
+- `semantic-color-surface-default-indicator` value
+- `semantic-color-surface-default-track` value
+- `semantic-color-surface-sub` value
+- `semantic-color-surface-chip` value
+- `semantic-color-surface-notification-card` value
+
+### Added
+
+- `color-semantic` tokens to match the latest guide
+- `radius-semantic` tokens to match the latest guide
+- `typography-semantic` tokens to match the latest guide
+
 ## [0.8.2] - 2025-09-09
 
 - No significant changes